feat: 日記投稿機能の作成
- Goの標準に従い，appディレクトリをinternalに名称変更
- 構造をシンプルにするために，ルートのmain.goに統合
	- （cmd/main.goと参照先が間違えていた）
- ヘルスチェックを追加
- Postmanによるテストで動作確認済み
